Changes on removal of Secret manager from operator
